Improve the task script for Torque pbsdsh tasks
Improve logic, wait conditions, sleep times and offsets to avoid race
conditions on the lock file. But allow jobs to quit early if the
simulation is already done, to avoid unnecessary delays or lock ups.
Previously, there was an infinite loop condition: individual jobs would
wait for a lock file forever, even if the simulation was done and the
lock file already deleted.

The script now creates an .egsjob file to record when job 1 (master)
starts and ends. Other jobs consult this file to check if the simulation
is done. Standard output logging is also improved with time stamps and
more detailed messages, for easier diagnosis when problems occur.

The run_pbsdsh_task script is used to launch individual jobs within a
distributed shell context (in the Torque scheduler).