Fix source collection and simple alias table
Fix phase-space file bug in the egs_source_collection, due to an
undefined value of setSimulationChunk. It has now been defined, similar
to what is done in egs_transformed_source. This bug implied that all
parallel run chunks for a simulation using a source collection and a
phsp file would start at the beginning of the file, resulting in extra
recycling of particles. This would have skewed the statistics and
introduced latent phase-space variance artifacts.

Fix a bug in EGS_SimpleAliasTable where certain sets of weights were not
sampled correctly. For example, using weight = 10 10 1 1 in a source
collection resulted in sampling ratios of 10:5.5:1:5.5 instead of
10:10:1:1. There was a bug in the algorithm, where a > comparison was
used instead of the >= operator.

Thanks to Patricia Oliver for reporting these issues.