Patch simple alias table algorithm
I don't agree with the previous fix. It solves the sample failure case
(10, 10, 1, 1), but this is coincidental and the code remains fragile,
because the there is no distinction between failing to find a high (or
low) bin, and exhausting the corresponding loop. Therefore excessive
aliasing might be performed, despite all bins being already equalized.
In the failure case this triggers an aliasing of the last bin with
itself. I patched this in a minimally disruptive way by inserting an
explicit index for the high (or low) bin that was found, and then
breaking away when no such bin is found (alias complete). There is also
no reference to the "i" index looping (n-1) times, which makes the
algorithm difficult to follow.

I would much prefer a cleaner, more explicit code, e.g.,

    // normalize distribution
    sum = 0.0;
    for (i=0; i<n; i++) sum += p[i];
    for (i=0; i<n; i++) p[i] /= sum;

    // sort into big and small bins
    for (i=0; i<n; i++) {
        q[i] = p[i]*n;
        if (q[i] < 1) small[nSmall++] = i;
        else big[nBig++] = i;
    }

    // aliasing
    while (nBig > 0) {
        s = small[(nSmall--)-1];
        b = big[nBig-1];
        bins[s] = b;
         q[b] = q[b] - (1-q[s]);
         if (q[b] < 1) {
             nBig--;
             small[nSmall++] = b;
         }
    }